requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
requests==2.32.3
psycopg[binary]==3.2.3
python-dotenv==1.0.1
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
